add prototype based cell height row builder
